chore(engine): combine ACTIVATING and ACTIVATED handler
* create child instance in the ACTIVATING handler to avoid problems with concurrent deployments
