+++ engine/src/main/java/io/zeebe/engine/processor/workflow/BpmnStepHandlers.java
@@ -16,7 +16,6 @@ import io.zeebe.engine.processor.workflow.handlers.activity.ActivityElementCompl
 import io.zeebe.engine.processor.workflow.handlers.activity.ActivityElementTerminatedHandler;
 import io.zeebe.engine.processor.workflow.handlers.activity.ActivityElementTerminatingHandler;
 import io.zeebe.engine.processor.workflow.handlers.activity.ActivityEventOccurredHandler;
-import io.zeebe.engine.processor.workflow.handlers.callactivity.CallActivityActivatedHandler;
 import io.zeebe.engine.processor.workflow.handlers.callactivity.CallActivityActivatingHandler;
 import io.zeebe.engine.processor.workflow.handlers.catchevent.IntermediateCatchEventElementActivatedHandler;
 import io.zeebe.engine.processor.workflow.handlers.catchevent.IntermediateCatchEventElementActivatingHandler;
@@ -172,10 +171,8 @@ public class BpmnStepHandlers {
         new MultiInstanceBodyEventOccurredHandler(stepHandlers::get));
 
     stepHandlers.put(
-        BpmnStep.CALL_ACTIVITY_ACTIVATING, new CallActivityActivatingHandler(catchEventSubscriber));
-    stepHandlers.put(
-        BpmnStep.CALL_ACTIVITY_ACTIVATED,
-        new CallActivityActivatedHandler(state.getKeyGenerator()));
+        BpmnStep.CALL_ACTIVITY_ACTIVATING,
+        new CallActivityActivatingHandler(catchEventSubscriber, state.getKeyGenerator()));
   }
 
   public void handle(final BpmnStepContext context) {
