+++ engine/src/main/java/io/zeebe/engine/processor/workflow/deployment/model/BpmnStep.java
@@ -66,6 +66,5 @@ public enum BpmnStep {
   MULTI_INSTANCE_TERMINATING,
   MULTI_INSTANCE_EVENT_OCCURRED,
 
-  CALL_ACTIVITY_ACTIVATING,
-  CALL_ACTIVITY_ACTIVATED
+  CALL_ACTIVITY_ACTIVATING
 }
